preparing the dual-hybrid to finish the upgrade
